web/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
dnspython==2.7.0
dnsrecon==0.10.1
dotenv==0.9.9
filelock==3.17.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
netaddr==1.3.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0
virtualenv==20.29.3
Werkzeug==3.1.3
